+++ examples/fractalMail/index.js
@@ -5,8 +5,9 @@ const io = require('socket.io-client')
 const realData = require('./data.service') // real data service
 const mockedData = require('./mocked-data.service') // mocked data service
 
-// realData | mockedData, for a real server uncomment socketio
+// realData | mockedData
 const data = realData
+
 let socket = io(data.serverName, {})
 
 data.connect(socket)