+++ examples/simpleRouter/modules/page1.js
@@ -6,16 +6,16 @@ const h = F.h
 module.exports = F.def({
   init: ({key}) => ({
     key,
-    color: 'red',
+    color: 'purple',
   }),
   inputs: {
     toggleColor: (ctx, Action, _) => Action.ToggleColor(),
   },
   actions: {
-    ToggleColor: [[], m => R.evolve({color: c => (c == 'pink') ? 'green' : 'red'}, m)],
+    ToggleColor: [[], m => R.evolve({color: c => (c == 'purple') ? 'green' : 'purple'}, m)],
   },
   interfaces: {
-    view: (ctx, i, m) => h('div', [
+    view: (ctx, i, m) => h('div', {key: m.key}, [
       h('div', 'Page 1 :)'),
       h('div', {
         style: {backgroundColor: m.color, color: 'white', padding: '20px'},