+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changes
 
-## V 0.0.7
+## v0.0.7
 
 - Add chat example
 - Add service to core
@@ -10,6 +10,17 @@
 - Engine exposes drivers and tasks
 - Improved chat example
 
-## V 0.0.8 (Working in)
+## v0.0.9-8
+
+- Working in new examples and documentation
+- Improved services
+
+## v0.0.10
 
 - Add Router module
+- Add free-style to core and some utilities
+- Fixed playground errors
+
+## v0.0.11 (working in)
+
+